organization: prepare for bringing in memmem
This basically bumps down all of the memchr tests/benchmarks into a
sub-module in order to make room for tests/benchmarks for other APIs
like memmem.
