+++ bench/src/data.rs
@@ -1,0 +1,6 @@
+pub const SHERLOCK_HUGE: &'static [u8] =
+    include_bytes!("../data/sherlock-holmes-huge.txt");
+pub const SHERLOCK_SMALL: &'static [u8] =
+    include_bytes!("../data/sherlock-holmes-small.txt");
+pub const SHERLOCK_TINY: &'static [u8] =
+    include_bytes!("../data/sherlock-holmes-tiny.txt");
